requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Others
python-dotenv==1.0.0
python-multipart==0.0.6
